feat: move structs module
