+++ backend/crates/talk-loco-client/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod client;
 pub mod macros;
-pub mod structs;
 pub mod talk;
 
 pub use futures_loco_protocol;
