--- backend/crates/talk-loco-client/src/structs/mod.rs
@@ -1,3 +1,0 @@
-pub mod channel;
-pub mod chat;
-pub mod openlink;
