+++ backend/crates/talk-loco-client/src/talk/mod.rs
@@ -1,2 +1,5 @@
+pub mod channel;
+pub mod chat;
+pub mod openlink;
 pub mod session;
 pub mod stream;
