+++ backend/crates/talk-loco-client/src/talk/session/channel/open/info.rs
@@ -1,13 +1,32 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-use crate::structs::{channel::ChannelMeta, chat::Chatlog, openlink::OpenLinkId};
+use crate::talk::{channel::ChannelMeta, chat::Chatlog, openlink::OpenLinkId};
 
 use super::user::DisplayUser;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(untagged)]
+pub enum ChannelType {
+    DirectChat,
+    MultiChat,
+
+    #[serde(rename = "OM")]
+    OpenMultiChat,
+
+    #[serde(rename = "OD")]
+    OpenDirectChat,
+
+    MemoChat,
+
+    PlusChat,
+
+    Other(String),
+}
 
 #[derive(Debug, Clone, Deserialize, PartialEq)]
 pub struct ChannelInfo {
     #[serde(rename = "type")]
-    pub channel_type: String,
+    pub channel_type: ChannelType,
 
     #[serde(flatten)]
     pub link: Option<OpenLinkId>,
